feat: package providing shared React context
