+++ packages/shared/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@hitachivantara/uikit-react-core",
-  "version": "5.0.0-next.13",
+  "name": "@hitachivantara/uikit-react-shared",
+  "version": "5.0.0-next.11",
   "private": false,
   "author": "Hitachi Vantara UI Kit Team",
   "description": "UI Kit core React components for the Next Design System.",
@@ -25,44 +25,20 @@
   "scripts": {
     "build": "run-s clean build:*",
     "build:js": "vite build",
-    "test": "vitest run",
-    "test:update": "vitest -u",
-    "test:ui": "vitest --ui",
-    "coverage": "vitest run --coverage",
     "clean": "npx rimraf dist package",
     "typecheck": "tsc --noEmit",
     "prepublishOnly": "pnpm build && npx clean-publish --fields main"
   },
   "peerDependencies": {
-    "@emotion/css": "^11.10.6",
-    "@emotion/react": "^11.10.5",
-    "@emotion/styled": "^11.10.5",
-    "@mui/material": "^5.11.0",
     "react": "^17.0.0 || ^18.0.0",
     "react-dom": "^17.0.0 || ^18.0.0"
   },
   "dependencies": {
-    "@emotion/cache": "^11.10.5",
-    "@hitachivantara/uikit-react-shared": "workspace:*",
-    "@hitachivantara/uikit-react-icons": "workspace:*",
-    "@hitachivantara/uikit-styles": "workspace:*",
-    "@popperjs/core": "^2.11.6",
-    "attr-accept": "^2.2.2",
-    "clsx": "^1.2.1",
-    "dayjs": "^1.11.7",
-    "detect-browser": "^5.3.0",
-    "hoist-non-react-statics": "^3.3.2",
-    "lodash": "^4.17.21",
-    "notistack": "^2.0.8",
-    "react-popper": "^2.3.0",
-    "react-resize-detector": "^8.0.4",
-    "react-table": "^7.8.0",
-    "react-window": "^1.8.8"
+    "@hitachivantara/uikit-styles": "workspace:*"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
-    "@types/react-dom": "^18.0.11",
-    "tsc-alias": "^1.8.2"
+    "@types/react-dom": "^18.0.11"
   },
   "files": [
     "dist"
@@ -75,9 +51,11 @@
     "types": "dist/types/index.d.ts",
     "exports": {
       ".": {
-        "require": "dist/cjs/index.cjs",
-        "import": "dist/esm/index.js"
-      }
+        "require": "./dist/cjs/index.cjs",
+        "import": "./dist/esm/index.js"
+      },
+      "./package.json": "./package.json",
+      "./bundles/*": "./dist/bundles/*"
     }
   },
   "clean-publish": {
