+++ packages/shared/vite.config.js
@@ -1,0 +1,17 @@
+import { mergeConfig } from "vite";
+
+import viteConfig from "../../.config/vite.config";
+
+export default mergeConfig(viteConfig, {
+  build: {
+    rollupOptions: {
+      output: [
+        {
+          format: "esm",
+          dir: "dist/bundles",
+          sourcemap: true,
+        },
+      ],
+    },
+  },
+});
